feat: upgrade to GO 1.27, new uuid type, etc. (#2)
* chore: bump GO v1.27, update UUID usage

* feat: remake uuid field

* fix: rename uuid GO type const

* fix: clean uuid field, update template for ID

* WIP: Apply new uuid flows over tests and regenerate subprojects, upgrade linter config to v2

* WIP: fix generation, update workflows, fix docker compose, add taskfile

* fix: convert uuid to string (mysql compat), update tests

* chore: fix tests golang matrix to 1.27 only for now

* fix: review feedback

+++ cmd/entfix/entfix.go
@@ -97,7 +97,7 @@ func (cmd *GlobalID) Run(ctx context.Context) error {
 	}
 	is := make(gen.IncrementStarts, len(ts))
 	for i, t := range ts {
-		is[t] = int64(i << 32)
+		is[t] = i << 32
 	}
 	if err := is.WriteToDisk(cmd.Path); err != nil {
 		return err
